feat: Enhance data display in commands and overview pages with additional event and voltage information. feat: Update DataServiceManager to include event states and voltage display logic. fix: Correct UART connection check in HM11 initialization. fix: Ensure proper handling of pyro firing status in Pyro_Fire function. refactor: Adjust system measurements to use uint16_t for voltage values. refactor: Improve event state management in ODB_Update function.

+++ ODB_V3/Software/STM32CubeIDE/ODBlix/Core/Inc/GAUL_Drivers/system_measurements.h
@@ -21,9 +21,9 @@ typedef struct {
 	uint16_t            pg_pin;
 
     float 				temperature;
-    float 				vin_batt;
-    float 				v5_buck;
-    float 				v3_buck;
+    uint16_t 			vin_batt;
+    uint16_t 			v5_buck;
+    uint16_t 			v3_buck;
     bool				pg_v5;
     bool 				pyros_arming;
     bool 				pyro_status[4];
